Refactor : 여정 조회 중 timeline 부분 수정
여정 조회 부분 중 timeline 부분 파일 이름 수정 및 mvvm 모델 적용
